[feat/45]: className 추가
